Querying through text field
Deleted logos and boiler plate react code. There might still be some old stuff lurking that needs to be changed. Added new logo file, created through Logo.js object. Added APIRecipe.js, which handles the API call. Added a text field and button that allows users to enter data and search. NOTE: the inbuilt set state function is taking a long time to update, might need to find a way around it. Right now you need to press the button twice for it to update, the results are displayed in the console.