Updated API call to display results
Had to mess around with the APIRecipe function/file to actually display the results, right now it displays them all under the header vertically. I kept the old function I made that turns the results into an array incase it's easier. The new function currently renders in the component when there isn't a NULL value. It still needs to be filled out to properly display all the values, and should probably be moved to a separate component.